+++ src/anvil_py.rs
@@ -1,9 +1,16 @@
 use url::{Url};
 
-use std::fs;
-use std::path::PathBuf;
+use std::fs::{self, File};
+use std::io::{self, Write};
+use std::path::{Path,PathBuf};
 use std::time::Instant;
 use toml::Table;
+use tar::Archive;
+use flate2::read::GzDecoder;
+use std::os::unix::fs::PermissionsExt;
+use std::fs::OpenOptions;
+
+use crate::core::{INVIL_VERSION, AmbosoEnv};
 
 pub const ANVILPY_PROJECT_VERSION_KEYNAME: &str = "version";
 pub const ANVILPY_PROJECT_NAME_KEYNAME: &str = "name";
@@ -288,4 +295,125 @@ fn parse_pyproject_tomlvalue(pyproj_str: &str, pyproj_path: &PathBuf, start_time
             return Err("Failed parsing TOML".to_string());
         }
     }
+}
+
+pub fn unpack_srcdist(pack_path: &PathBuf) -> Result<PathBuf,String> {
+    debug!("TODO: add unpack step");
+    let tar_gz = File::open(pack_path);
+    match tar_gz {
+        Ok(tar_gz) => {
+            let tar = GzDecoder::new(tar_gz);
+            let mut archive = Archive::new(tar);
+            let mut outdir = pack_path.clone();
+            outdir.pop();
+            let unpack_res = archive.unpack(outdir.clone());
+            match unpack_res {
+                Ok(_) => {
+                    debug!("Unpacked srcdist {{{}}} to {{{}}}", pack_path.display(), outdir.display());
+                    return Ok(outdir);
+                }
+                Err(e) => {
+                    error!("Failed unpacking srcdist from {{{}}}. Err: {e}", pack_path.display());
+                    return Err("Failed unpacking srcdist".to_string());
+                }
+            }
+        }
+        Err(e) => {
+            error!("Failed opening srcdist pack at {{{}}}. Err: {e}", pack_path.display());
+            return Err("Failed opening srcdist".to_string());
+        }
+    }
+}
+
+
+pub fn gen_anvilpy_shim(shim_path: &PathBuf, module_path: &str, entrypoint_func: &str) -> Result<String,String> {
+    trace!("Generating anvilpy shim. Target path: {{{}}} Module path: {{{}}} Function: {{{}}}", shim_path.display(), module_path, entrypoint_func);
+    let output = File::create(shim_path);
+    let shim_string = format!("#!/bin/python3\n\n##\n# Generated by invil v{INVIL_VERSION}\n# Repo at https://github.com/jgabaut/invil\n##\n\nimport sys\nimport re\nfrom {ANVILPY_UNPACKDIR_NAME}.{module_path} import {entrypoint_func}\n\nif __name__ == \'__main__\':\n    sys.argv[0] = re.sub(r'(-script\\.pyw|\\.exe)?$', '', sys.argv[0])\n    sys.exit({entrypoint_func}())\n");
+    match output {
+        Ok(mut f) => {
+            let res = write!(f, "{}", shim_string);
+            match res {
+                Ok(_) => {
+                    debug!("Done generating shim file");
+                }
+                Err(e) => {
+                    error!("Failed printing shim file");
+                    return Err(e.to_string());
+                }
+            }
+            if cfg!(target_os = "windows") {
+                todo!("Support setting executable bit for shim on windows");
+                /*
+                 * let output = Command::new("cmd")
+                 *   .args(["/C", "echo hello"])
+                 *   .output()
+                 *   .expect("failed to execute process")
+                 */
+            } else {
+                let metadata = f.metadata();
+                match metadata {
+                    Ok(m) => {
+                        trace!("Setting mode 744 for shim");
+                        let mut permissions = m.permissions();
+                        // Add execute permission for the owner
+                        permissions.set_mode(permissions.mode() | 0o100);
+                        // Set the updated permissions
+                        fs::set_permissions(shim_path, permissions).expect("Failed to set file permissions");
+                    }
+                    Err(e) => {
+                        error!("Failed getting metadata for shim. Err: {e}");
+                        return Err("Failed getting shim metadata".to_string());
+                    }
+                }
+            }
+            return Ok(format!("Generated {{{}}}", shim_path.display()));
+        }
+        Err(_) => {
+            return Err("Failed gen of shim file".to_string());
+        }
+    }
+
+}
+
+// From https://doc.rust-lang.org/rust-by-example/std_misc/fs.html
+// A simple implementation of `% touch path` (ignores existing files)
+fn touch_file(path: &Path) -> io::Result<()> {
+    match OpenOptions::new().create(true).write(true).open(path) {
+        Ok(_) => Ok(()),
+        Err(e) => Err(e),
+    }
+}
+
+pub fn post_unpack(unpack_initpy_path: &PathBuf, bindir_path: &PathBuf, env: &AmbosoEnv) -> Result<String,String> {
+    match touch_file(&unpack_initpy_path) {
+        Ok(_) => {},
+        Err(e) => {
+            error!("Failed touch for {{{}}}. Err: {e}", unpack_initpy_path.display());
+            return Err("Failed touch for unpack initpy".to_string());
+        }
+    }
+
+    debug!("TODO: prep shims for all entrypoints");
+
+    let mut shims_made = 0;
+    for entry in &env.anvilpy_env.as_ref().expect("Failed initialising anvilpy_env").scripts {
+        let entrypoint_parts: Vec<&str> = entry.entrypoint.splitn(2, ':').collect();
+        let module_path = entrypoint_parts[0];
+        let entrypoint_funcname = entrypoint_parts[1];
+        let shimname = &entry.name;
+        debug!("TODO: prep shim {{{}}} -> {{{}}} : {{{}}}", shimname, module_path, entrypoint_funcname);
+        let mut shim_path = bindir_path.clone();
+        shim_path.push(shimname);
+        let shimgen_res = gen_anvilpy_shim(&shim_path, module_path, entrypoint_funcname);
+        match shimgen_res {
+            Ok(_) => { shims_made += 1; }
+            Err(e) => {
+                error!("Failed generating shim at {{{}}}", shim_path.display());
+                return Err(e);
+            }
+        }
+    }
+
+    Ok(format!("Shims made: {{{shims_made}}}"))
 }
